+++ src/app/app.component.ts
@@ -1,17 +1,16 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 import { Platform, MenuController } from '@ionic/angular';
 import { SplashScreen } from '@ionic-native/splash-screen/ngx';
 import { StatusBar } from '@ionic-native/status-bar/ngx';
 import { TranslateConfigService } from './services/translate-config.service';
-import { AppVersion } from '@ionic-native/app-version/ngx';
 
 @Component({
   selector: 'app-root',
   templateUrl: 'app.component.html',
   styleUrls: ['app.component.scss']
 })
-export class AppComponent implements OnInit {
+export class AppComponent {
   public appPages = [
     {
       title: 'MENU.home',
@@ -20,8 +19,8 @@ export class AppComponent implements OnInit {
     },
     {
       title: 'MENU.models-comparison',
-      url: '/all-models',
-      icon: 'checkmark-circle'
+      url: '/models/all-models',
+      icon: 'checkmark'
     },
     {
       title: 'MENU.favorites',
@@ -32,27 +31,29 @@ export class AppComponent implements OnInit {
       title: 'MENU.saved-searches',
       url: '/searches',
       icon: 'bookmark'
+    },
+    {
+      title: 'MENU.comparisons',
+      url: '/comparisons',
+      icon: 'git-compare'
     }
   ];
-  public version: string;
 
   constructor(
     private platform: Platform,
     private splashScreen: SplashScreen,
     private statusBar: StatusBar,
     private translationConfigService: TranslateConfigService,
-    private menuController: MenuController,
-    private appVersion: AppVersion
+    private menuController: MenuController
   ) {
     this.initializeApp();
-    this.version = '';
   }
 
   private initializeApp(): void {
     this.platform.ready().then(() => {
       this.statusBar.overlaysWebView(false);
-      this.statusBar.backgroundColorByHexString('#FFFFFF');
-      this.statusBar.styleDefault();
+      this.statusBar.backgroundColorByHexString('#222428');
+      this.statusBar.styleBlackOpaque();
 
       // Translate
       this.translationConfigService.setTranslation();
@@ -63,12 +64,5 @@ export class AppComponent implements OnInit {
 
   public async closeMenu(): Promise<boolean> {
     return this.menuController.close();
-  }
-
-  public async ngOnInit(): Promise<void> {
-    // Get app version
-    const versionCode = await this.appVersion.getVersionCode();
-    const versionNumber = await this.appVersion.getVersionNumber();
-    this.version = `${versionNumber} - build ${versionCode}`;
   }
 }