+++ src/app/interfaces/mockup-logos.ts
@@ -3,3 +3,10 @@ export const AUDI = 'https://pluspng.com/img-png/audi-logo-png-1920x1080-hd-1080
 export const KIA = 'https://i0.wp.com/www.maggioraoffroadarena.it/wp-content/uploads/2019/08/kia-logo.png?fit=756%2C387&ssl=1';
 export const OPEL = 'https://www.cavallomotorigroup.it/wp-content/uploads/2018/09/Opel-logo-2002-2048x2048.png';
 export const SEAT = 'https://www.metromaffesa.com/wp-content/uploads/2018/01/adhesivo-seat-logo-efecto-3d-impreso.jpg.png';
+
+export const A585 = 'https://www.abarth.it/content/dam/abarth/gamma/range2017/Car_Slider_595_640x343-01.png';
+export const A3 = 'https://immagini.alvolante.it/sites/default/files/styles/anteprima_lunghezza_640_jpg/public/serie_auto_galleria/2012/10/audi_a3_sportback_ant_0.png?itok=LsX22GTR';
+export const A4 = 'https://immagini.alvolante.it/sites/default/files/styles/anteprima_lunghezza_640_jpg/public/serie_auto_galleria/2015/09/audi_a4_ant_1.png?itok=Z6p1mUyT';
+export const RIO = 'https://www.kia.com/content/dam/kwcms/kme/global/en/assets/vehicles/rio-YB/discover/kia-rio-yb-my19-trim-ex.png';
+export const CORSA = 'https://immagini.alvolante.it/sites/default/files/styles/image_gallery_big/public/serie_auto_galleria/2019/07/opel_corsa_ant.png?itok=U8B9PNLC';
+export const IBIZA = 'https://immagini.alvolante.it/sites/default/files/styles/anteprima_lunghezza_640/public/serie_auto_galleria/2017/04/seat_ibiza_ant_0.png';