+++ src/app/modules/item-model/item-model.component.ts
@@ -1,0 +1,17 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'app-item-model',
+  templateUrl: './item-model.component.html',
+  styleUrls: ['./item-model.component.scss'],
+})
+export class ItemModelComponent {
+
+  @Input() title: string;
+  @Input() img: string;
+  @Input() manufacturer: string;
+  @Input() fuel: string;
+
+  constructor() { }
+
+}