+++ src/app/modules/models-module/models/models.page.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import { AbstractBackNavigationPage } from '../../utils/abstract-back-navigation';
-import { BackNavigationService } from '../../services/back-navigation.service';
+import { AbstractBackNavigationPage } from '../../../utils/abstract-back-navigation';
+import { BackNavigationService } from '../../../services/back-navigation.service';
 import { DataFetcherService } from 'src/app/services/data-fetcher.service';
+import { IModel } from 'src/app/interfaces/model.interface';
 
 @Component({
   selector: 'app-models',
@@ -11,6 +12,7 @@ import { DataFetcherService } from 'src/app/services/data-fetcher.service';
 export class ModelsPage extends AbstractBackNavigationPage implements OnInit {
 
   public selectedManufacturer: string;
+  public models: IModel[];
 
   constructor(
     backNavigationService: BackNavigationService,
@@ -18,10 +20,12 @@ export class ModelsPage extends AbstractBackNavigationPage implements OnInit {
   ) {
     super(backNavigationService, { toHome: false, inRoot: false });
     this.selectedManufacturer = '';
+    this.models = [];
   }
 
   public ngOnInit() {
     this.selectedManufacturer = this.dataFetchService.getSelectedManufacturer();
+    this.models = this.dataFetchService.getModelsByManufacturer(this.selectedManufacturer);
   }
 
 }